progress towards #53 . VSTS components have been refactored, although approach needs discussion with the team
